User friendly improve

+++ gwt-ol-main/src/main/java/org/openlayers/ol/style/options/StrokeStyleOptions.java
@@ -71,7 +71,7 @@ public class StrokeStyleOptions {
      * Width.
      */
     @JsProperty
-    public double width;
+    public int width;
 
 
     @JsConstructor
